+++ examples/pipeline/intermediate/pipeline.py
@@ -5,11 +5,12 @@
 # Note: run this using `ipython` or in a Jupyter notebook (it won't run using `python`).
 
 # +
+from pathlib import Path
 import logging
 
 from ploomber.products import File, PostgresRelation
-from ploomber.tasks import (BashCommand, PythonCallable,
-                                    SQLScript)
+from ploomber.tasks import (ShellScript, PythonCallable,
+                            SQLScript)
 from ploomber.dag import DAG
 from ploomber.clients import SQLAlchemyClient
 from ploomber import testing
@@ -28,8 +29,7 @@ logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
 
 
-env = Env()
-home = env.path.home
+env = Env.start()
 path_to_sample = env.path.input / 'sample'
 
 (env.path.input / 'raw').mkdir(exist_ok=True, parents=True)
@@ -61,7 +61,7 @@ dag.product.delete()
 # let's introduce a new Task: bash command, this command will generate
 # 3 products, it is recommended for Tasks to have a single Product
 # but you can add more if needed
-get_data = BashCommand((home / 'get_data.sh').read_text(),
+get_data = ShellScript(Path('get_data.sh').read_text(),
                        (File(env.path.input / 'raw' / 'red.csv'),
                         File(env.path.input / 'raw' / 'white.csv'),
                         File(env.path.input / 'raw' / 'names')),
@@ -79,16 +79,15 @@ get_data >> sample
 # utility, furthermore, since we declared a product and an upstream dependency, those are
 # available as parameters as well. This task does not create a File but a PostgresRelation,
 # this type of product is a a table named red in the public schema
-red_task = BashCommand(('csvsql --db {{uri}} --tables {{product.name}} --insert {{upstream["sample"][0]}} '
+red_task = ShellScript(('csvsql --db {{uri}} --tables {{product.name}} --insert {{upstream["sample"][0]}} '
                         '--overwrite'),
                        PostgresRelation(('public', 'red', 'table')),
                        dag,
                        params=dict(uri=uri),
-                       split_source_code=False,
                        name='red')
 sample >> red_task
 
-white_task = BashCommand(('csvsql --db {{uri}} --tables {{product.name}} --insert {{upstream["sample"][1]}} '
+white_task = ShellScript(('csvsql --db {{uri}} --tables {{product.name}} --insert {{upstream["sample"][1]}} '
                           '--overwrite'),
                          PostgresRelation(('public', 'white', 'table')),
                          dag,
@@ -99,19 +98,19 @@ sample >> white_task
 
 # let's introduce a new type of task, a SQLScript, this task will execute
 # a script in a SQL database (any database supported by sql alchemy)
-wine_task = SQLScript(home / 'sql' / 'create_wine.sql',
+wine_task = SQLScript(Path('sql', 'create_wine.sql'),
                       PostgresRelation(('public', 'wine', 'table')),
                       dag, name='wine')
 (red_task + white_task) >> wine_task
 
 
-dataset_task = SQLScript(home / 'sql' / 'create_dataset.sql',
+dataset_task = SQLScript(Path('sql', 'create_dataset.sql'),
                          PostgresRelation(('public', 'dataset', 'table')),
                          dag, name='dataset')
 wine_task >> dataset_task
 
 
-training_task = SQLScript(home / 'sql' / 'create_training.sql',
+training_task = SQLScript(Path('sql', 'create_training.sql'),
                           PostgresRelation(('public', 'training', 'table')),
                           dag, name='training')
 dataset_task >> training_task
@@ -119,7 +118,7 @@ dataset_task >> training_task
 
 testing_table = PostgresRelation(('public', 'testing', 'table'))
 # testing_table.tests = [testing.Postgres.no_nas_in_column('label')]
-testing_task = SQLScript(home / 'sql' / 'create_testing.sql',
+testing_task = SQLScript(Path('sql', 'create_testing.sql'),
                          testing_table, dag, name='testing')
 
 dataset_task >> testing_task