+++ src/ploomber/env/__init__.py
@@ -1,3 +1,0 @@
-from ploomber.env.env import Env
-
-__all__ = ['Env']