+++ src/ploomber/templates/__init__.py
@@ -1,3 +1,0 @@
-from ploomber.templates.SQLStore import SQLStore
-
-__all__ = ['SQLStore']