+++ src/ploomber/validators/validators.py
@@ -40,7 +40,7 @@ class Assert:
         elif len(self.messages_warning) > 1:
             str_ += ('\n\n {} warnings: \n * {}'
                      .format(len(self.messages_warning),
-                             '\n * '.join(self.messages_error)))
+                             '\n * '.join(self.messages_warning)))
 
         return str_
 
@@ -73,10 +73,11 @@ def validate_schema(assert_, data, schema, error_on_extra_cols=False):
     missing = expected - cols
     unexpected = cols - expected
 
-    msg = 'Missing columns: {missing}.'.format(missing=missing)
+    msg = 'validate_schema: missing columns {missing}.'.format(missing=missing)
     assert_(not missing, msg)
 
-    msg = 'Unexpected columns {unexpected}'.format(unexpected=unexpected)
+    msg = ('validate_schema: wnexpected columns {unexpected}'
+           .format(unexpected=unexpected))
     caller = assert_ if error_on_extra_cols else assert_.warn
     caller(not unexpected, msg)
 
@@ -87,7 +88,7 @@ def validate_schema(assert_, data, schema, error_on_extra_cols=False):
         expected = schema.get(name)
 
         if expected is not None:
-            msg = ('Wrong dtype for column "{name}". '
+            msg = ('validate_schema: wrong dtype for column "{name}". '
                    'Expected: "{expected}". Got: "{dtype}"'
                    .format(name=name, expected=expected, dtype=dtype))
             assert_(dtype == expected, msg)
@@ -95,16 +96,54 @@ def validate_schema(assert_, data, schema, error_on_extra_cols=False):
     return assert_
 
 
+@validator
+def validate_values(assert_, data, values):
+    data_cols = data.columns
+
+    for column, (kind, params) in values.items():
+        if column not in data_cols:
+            assert_.warn(False,
+                         ('validate_values: declared spec for column "{}" but'
+                          ' it does not appear in the data').format(column))
+        elif kind == 'unique':
+            expected = set(params)
+            unique = set(data[column].unique())
+            msg = ('validate_values:: expected unique values of  "{}" to be a'
+                   ' subset of {}, got: {}'
+                   .format(column, expected, unique))
+            assert_(expected >= unique, msg)
+        elif kind == 'range':
+            if len(params) != 2:
+                raise ValueError('If kind is range, you must provide two '
+                                 'values, got {}'.format(params))
+            min_expected, max_expected = params
+            min_ = data[column].min()
+            max_ = data[column].max()
+            msg = ('validate_values: expected range of "{}" to be ({}, {}), '
+                   'got ({}, {})'
+                   .format(column, min_expected, max_expected, min_, max_))
+            assert_(min_expected <= min_ and max_ <= max_expected, msg)
+        else:
+            ValueError('Got invalid kind, must be "unique" or "range"')
+
+
 def data_frame_validator(df, validators):
     """
 
     Examples
     --------
+    >>> from ploomber.validators import data_frame_validator
+    >>> from ploomber.validators import validate_schema, validate_values
     >>> import pandas as pd
     >>> import numpy as np
-    >>> df = pd.DataFrame({'x': np.random.rand(10), 'y': np.random.rand(10)})
+    >>> df = pd.DataFrame({'x': np.random.rand(3), 'y': np.random.rand(3),
+    ...                    'z': [0, 1, 2], 'i': ['a', 'b', 'c']})
     >>> data_frame_validator(df,
-    ...                     [validate_schema(schema={'x': 'int', 'z': 'int'})])
+    ...                     [validate_schema(schema={'x': 'int', 'z': 'int'}),
+    ...                      validate_values(values={'z': ('range', (0, 1)),
+    ...                                              'i': ('unique', {'a'}),
+    ...                                              'j': ('unique', {'b'})}
+    ...                                      )])
     """
     assert_ = Assert()