+++ tests/test_env.py
@@ -2,8 +2,23 @@ import pytest
 from ploomber.env.env import _get_name, Env
 
 
+def test_init_with_module_key(cleanup_env):
+    env = Env.start({'module': 'sample_project'})
+    assert env.module == 'sample_project'
+
+
+def test_init_with_nonexistent_package(cleanup_env):
+    with pytest.raises(ImportError):
+        Env.start({'module': 'i_do_not_exist'})
+
+
+def test_version_placeholder(cleanup_env):
+    env = Env.start({'module': 'sample_project', 'version': '{{version}}'})
+    assert env.version == '0.1'
+
+
 def test_can_create_env_from_dict(cleanup_env):
-    e = Env.from_dict({'a': 1})
+    e = Env.start({'a': 1})
     assert e.a == 1
 
 
@@ -22,9 +37,9 @@ def test_raises_error_if_wrong_format():
 
 def test_can_instantiate_env_if_located_in_sample_dir(move_to_sample_dir,
                                                       cleanup_env):
-    Env()
+    Env.start()
 
 
 def test_can_instantiate_env_if_located_in_sample_subdir(move_to_sample_subdir,
                                                          cleanup_env):
-    Env()
+    Env.start()